+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,7 @@
 
 -->
 ## Upcoming Release
+* Added parameter `-EnableAutomaticUpgrade` to `Set-AzVmExtension` and `Add-AzVmssExtension`.
 * Removed FilterExpression parameter from `Get-AzVMImage` cmdlet documentation. 
 * Added deprecation message to the ContainerService cmdlets:
     - `Add-AzureRmContainerServiceAgentPoolProfileCommand`